gantt refactor and docu
