+++ src/management-system-v2/components/bpmn-timeline/transform.ts
@@ -17,13 +17,18 @@ import type {
   DEFAULT_DURATIONS
 } from './types';
 import { calculatePathBasedTimings } from './path-traversal';
+import { calculateElementTimings } from './timing-calculator';
+import {
+  transformTask,
+  transformEvent,
+  transformSequenceFlow,
+  getFlowType,
+} from './element-transformers';
 import {
   extractDuration,
   isTaskElement,
   isSupportedEventElement,
   isSequenceFlowElement,
-  getTaskTypeString,
-  getEventTypeString,
   getUnsupportedElementReason,
   formatElementForLog,
   formatTimingForLog,
@@ -35,301 +40,6 @@ import {
 } from './utils';
 
 // ============================================================================
-// Element Transformation Functions
-// ============================================================================
-
-/**
- * Transform BPMN task to Gantt task
- */
-function transformTask(
-  task: BPMNTask, 
-  startTime: number, 
-  duration: number,
-  color?: string
-): GanttElementType {
-  return {
-    id: task.id,
-    name: task.name,
-    type: 'task',
-    start: startTime,
-    end: startTime + duration,
-    elementType: getTaskTypeString(task),
-    color
-  };
-}
-
-/**
- * Transform BPMN event to Gantt milestone
- * Using Option A: Events positioned at the end of their duration (completion time)
- */
-function transformEvent(
-  event: BPMNEvent,
-  startTime: number,
-  duration: number,
-  color?: string
-): GanttElementType {
-  if (duration > 0) {
-    // Event has duration - show the time range with milestone at completion time
-    // Option A: milestone appears at event completion time but shows the full range
-    return {
-      id: event.id,
-      name: event.name,
-      type: 'milestone',
-      start: startTime,
-      end: startTime + duration,
-      elementType: getEventTypeString(event),
-      color
-    };
-  } else {
-    // Event has no duration - point milestone at start time
-    return {
-      id: event.id,
-      name: event.name,
-      type: 'milestone',
-      start: startTime,
-      elementType: getEventTypeString(event),
-      color
-    };
-  }
-}
-
-/**
- * Determine flow type based on BPMN properties
- */
-function getFlowType(flow: BPMNSequenceFlow): 'conditional' | 'default' | 'normal' {
-  // Check if flow has a condition expression (conditional flow)
-  if (flow.conditionExpression) {
-    return 'conditional';
-  }
-  
-  // Check if flow is marked as default (from gateway default sequence flow)
-  // In BPMN, default flows are typically marked on the gateway's default property
-  // but we can also check for specific naming patterns or attributes
-  if (flow.name && (flow.name.toLowerCase().includes('default') || flow.name.toLowerCase().includes('else'))) {
-    return 'default';
-  }
-  
-  return 'normal';
-}
-
-/**
- * Transform BPMN sequence flow to Gantt dependency
- */
-function transformSequenceFlow(flow: BPMNSequenceFlow): GanttDependency {
-  // Handle case where sourceRef/targetRef might be ModdleElement objects or strings
-  const sourceId = typeof flow.sourceRef === 'string' ? flow.sourceRef : (flow.sourceRef as any)?.id || flow.sourceRef;
-  const targetId = typeof flow.targetRef === 'string' ? flow.targetRef : (flow.targetRef as any)?.id || flow.targetRef;
-  
-  return {
-    id: flow.id,
-    sourceId: sourceId,
-    targetId: targetId,
-    type: DependencyType.FINISH_TO_START, // BPMN sequence flows are finish-to-start dependencies
-    name: flow.name,
-    flowType: getFlowType(flow)
-  };
-}
-
-// ============================================================================
-// Time Calculation Algorithm
-// ============================================================================
-
-/**
- * Calculate element timings using improved First Possible Occurrence algorithm
- * This algorithm properly handles early occurrence updates when later flows enable earlier start times
- */
-export function calculateElementTimings(
-  elements: BPMNFlowElement[],
-  startTime: number,
-  defaultDurations: DefaultDurationInfo[] = []
-): Map<string, ElementTiming> {
-  const timings = new Map<string, ElementTiming>();
-  const elementMap = new Map<string, BPMNFlowElement>();
-  const incomingFlows = new Map<string, string[]>();
-  const outgoingFlows = new Map<string, string[]>();
-  const trackedDefaultDurations = new Set<string>(); // Track which elements already have default durations recorded
-  
-  // Build element lookup and flow maps
-  elements.forEach(element => {
-    elementMap.set(element.id, element);
-    
-    if (element.$type === 'bpmn:SequenceFlow') {
-      const flow = element as BPMNSequenceFlow;
-      
-      // Extract actual IDs from sourceRef/targetRef (might be ModdleElements)
-      const sourceId = typeof flow.sourceRef === 'string' ? flow.sourceRef : (flow.sourceRef as any)?.id || flow.sourceRef;
-      const targetId = typeof flow.targetRef === 'string' ? flow.targetRef : (flow.targetRef as any)?.id || flow.targetRef;
-      
-      // Track incoming flows for target
-      const targetIncoming = incomingFlows.get(targetId) || [];
-      targetIncoming.push(flow.id);
-      incomingFlows.set(targetId, targetIncoming);
-      
-      // Track outgoing flows for source
-      const sourceOutgoing = outgoingFlows.get(sourceId) || [];
-      sourceOutgoing.push(flow.id);
-      outgoingFlows.set(sourceId, sourceOutgoing);
-    }
-  });
-  
-  // Get all non-flow elements
-  const allElements = elements.filter(el => el.$type !== 'bpmn:SequenceFlow');
-  
-  // Initialize timings for elements without incoming flows
-  allElements.forEach(element => {
-    const elementIncoming = incomingFlows.get(element.id) || [];
-    if (elementIncoming.length === 0) {
-      // Element has no incoming flows - starts at startTime
-      let duration = extractDuration(element);
-      if (duration === 0) {
-        // Apply default duration based on element type
-        if (isTaskElement(element)) {
-          duration = 3600000; // 1 hour default for tasks
-          if (!trackedDefaultDurations.has(element.id)) {
-            defaultDurations.push({
-              elementId: element.id,
-              elementType: element.$type,
-              elementName: element.name,
-              appliedDuration: duration,
-              durationType: 'task'
-            });
-            trackedDefaultDurations.add(element.id);
-          }
-        } else if (isSupportedEventElement(element)) {
-          duration = 0; // 0ms default for events
-          // Don't track default durations for events since 0ms is expected
-        }
-      }
-      
-      timings.set(element.id, {
-        elementId: element.id,
-        startTime: startTime,
-        endTime: startTime + duration,
-        duration
-      });
-    }
-  });
-  
-  // Iteratively calculate timings until no more updates occur
-  const maxIterations = allElements.length * 2; // Safety limit
-  let iteration = 0;
-  let hasUpdates = true;
-  
-  while (hasUpdates && iteration < maxIterations) {
-    hasUpdates = false;
-    iteration++;
-    
-    // Process each element that has incoming flows
-    allElements.forEach(element => {
-      const elementId = element.id;
-      const elementIncoming = incomingFlows.get(elementId) || [];
-      
-      if (elementIncoming.length === 0) {
-        return; // Already handled in initialization
-      }
-      
-      // Calculate the earliest possible start time from all incoming flows
-      let earliestStartTime = Number.MAX_SAFE_INTEGER;
-      let hasValidIncoming = false;
-      
-      for (const flowId of elementIncoming) {
-        const flow = elementMap.get(flowId) as BPMNSequenceFlow;
-        if (!flow) continue;
-        
-        // Extract source ID (might be ModdleElement)
-        const sourceId = typeof flow.sourceRef === 'string' ? flow.sourceRef : (flow.sourceRef as any)?.id || flow.sourceRef;
-        const sourceTiming = timings.get(sourceId);
-        
-        if (sourceTiming) {
-          // Add flow duration to source completion time
-          const flowDuration = extractDuration(flow) || 0;
-          const flowCompletionTime = sourceTiming.endTime + flowDuration;
-          earliestStartTime = Math.min(earliestStartTime, flowCompletionTime);
-          hasValidIncoming = true;
-        }
-      }
-      
-      if (!hasValidIncoming) {
-        return; // Not all dependencies are ready yet
-      }
-      
-      // Calculate duration for this element
-      let duration = extractDuration(element);
-      if (duration === 0) {
-        // Apply default duration based on element type
-        if (isTaskElement(element)) {
-          duration = 3600000; // 1 hour default for tasks
-          if (!trackedDefaultDurations.has(element.id)) {
-            defaultDurations.push({
-              elementId: element.id,
-              elementType: element.$type,
-              elementName: element.name,
-              appliedDuration: duration,
-              durationType: 'task'
-            });
-            trackedDefaultDurations.add(element.id);
-          }
-        } else if (isSupportedEventElement(element)) {
-          duration = 0; // 0ms default for events
-          // Don't track default durations for events since 0ms is expected
-        }
-      }
-      
-      const newEndTime = earliestStartTime + duration;
-      const existingTiming = timings.get(elementId);
-      
-      // Check if this is a new timing or an improvement (earlier start time)
-      if (!existingTiming || earliestStartTime < existingTiming.startTime) {
-        // Update timing - this is either new or an early occurrence update
-        timings.set(elementId, {
-          elementId,
-          startTime: earliestStartTime,
-          endTime: newEndTime,
-          duration
-        });
-        
-        hasUpdates = true; // Signal that we made an update
-      }
-    });
-  }
-  
-  // Handle any remaining unprocessed elements (likely in cycles or missing dependencies)
-  allElements.forEach(element => {
-    if (!timings.has(element.id)) {
-      let duration = extractDuration(element);
-      if (duration === 0) {
-        if (isTaskElement(element)) {
-          duration = 3600000; // 1 hour for tasks
-          if (!trackedDefaultDurations.has(element.id)) {
-            defaultDurations.push({
-              elementId: element.id,
-              elementType: element.$type,
-              elementName: element.name,
-              appliedDuration: duration,
-              durationType: 'task'
-            });
-            trackedDefaultDurations.add(element.id);
-          }
-        } else {
-          duration = 0; // 0 for events
-          // Don't track default durations for events since 0ms is expected
-        }
-      }
-      
-      // For elements with unresolvable dependencies, place them at start time
-      timings.set(element.id, {
-        elementId: element.id,
-        startTime: startTime,
-        endTime: startTime + duration,
-        duration
-      });
-      
-    }
-  });
-  
-  
-  return timings;
-}
 
 // ============================================================================
 // Main Transformation Function
